// src/logistics-api/model/createDriver.ts
/**
 * Logistics API
 * Logistics API provides services to find and manage entities for logistics
 *
 * OpenAPI spec version: 1.0
 * Contact: vladislavprokopenko1@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Object representing Driver creating
 */
export interface CreateDriver { 
    driverAuthenticationId?: number;
    name?: string;
    surname?: string;
    phone?: string;
    email?: string;
    workExperience?: number;
    workingHoursInCurrentMonth?: number;
    currentCity?: string;
    currentState?: string;
}